# server/docs/swagger.yaml
basePath: /
definitions:
  models.AuthResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/models.UserResponse'
    type: object
  models.DocumentDownloadURLResponse:
    properties:
      expires_in:
        type: integer
      url:
        type: string
    type: object
  models.DocumentResponse:
    properties:
      authors:
        type: string
      category:
        type: string
      created_at:
        type: string
      file_path:
        type: string
      id:
        type: string
      status:
        type: string
      title:
        type: string
      year:
        type: integer
    type: object
  models.DocumentStatusResponse:
    properties:
      status:
        description: pending, processing, done, error
        type: string
    type: object
  models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  models.RegisterRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  models.SearchResponse:
    properties:
      results:
        items:
          $ref: '#/definitions/models.SearchResultItem'
        type: array
    type: object
  models.SearchResultItem:
    properties:
      answer:
        type: string
      chunk_id:
        type: string
      confidence:
        type: number
      content:
        type: string
      document_id:
        type: string
    type: object
  models.UploadResponse:
    properties:
      document_id:
        type: string
      status:
        type: string
    type: object
  models.UserResponse:
    properties:
      email:
        type: string
      id:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: arbon
  description: API для интеллектуального поиска по документам.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: PDF Search API
  version: "1.0"
paths:
  /documents:
    get:
      description: Возвращает метаданные всех загруженных документов.
      parameters:
      - description: Максимальное количество документов на странице (по умолчанию
          20, макс 100)
        in: query
        name: limit
        type: integer
      - description: Смещение от начала списка (по умолчанию 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.DocumentResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Failed to fetch documents
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Список документов
      tags:
      - documents
  /documents/{id}:
    delete:
      description: Удаляет документ по ID и его PDF-файл.
      parameters:
      - description: ID документа
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid document ID
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Document not found
          schema:
            type: string
        "500":
          description: Failed to delete file or database record
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Удалить документ
      tags:
      - documents
    get:
      description: Возвращает метаданные документа по ID.
      parameters:
      - description: ID документа
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DocumentResponse'
        "400":
          description: Invalid document ID
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Document not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Получить документ
      tags:
      - documents
  /documents/{id}/download:
    get:
      description: Возвращает PDF-файл документа (редирект на временную ссылку MinIO)
      parameters:
      - description: ID документа
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/pdf
      responses:
        "302":
          description: Redirect to presigned URL
          schema:
            type: string
        "400":
          description: Invalid document ID
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Document not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Скачать документ
      tags:
      - documents
  /documents/{id}/download-url:
    get:
      description: Возвращает presigned URL для доступа к файлу документа.
      parameters:
      - description: ID документа
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DocumentDownloadURLResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Document not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Получить ссылку на скачивание
      tags:
      - documents
  /documents/{id}/status:
    get:
      description: Возвращает статус (pending, processing, done, error) документа
        по ID.
      parameters:
      - description: ID документа
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DocumentStatusResponse'
        "400":
          description: Invalid document ID
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Document not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Получить статус документа
      tags:
      - documents
  /login:
    post:
      consumes:
      - application/json
      description: Аутентификация пользователя, получение JWT.
      parameters:
      - description: Учётные данные
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Вход в систему
      tags:
      - auth
  /register:
    post:
      consumes:
      - application/json
      description: Создаёт нового пользователя с email и паролем.
      parameters:
      - description: Данные для регистрации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Регистрация пользователя
      tags:
      - auth
  /search:
    get:
      description: Полнотекстовый или семантический поиск по содержимому.
      parameters:
      - description: Поисковый запрос
        in: query
        name: q
        required: true
        type: string
      - description: 'Тип поиска: text (по умолчанию) или vector/semantic'
        in: query
        name: type
        type: string
      - description: Максимальное количество результатов (макс 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SearchResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Поиск документов
      tags:
      - search
  /upload:
    post:
      consumes:
      - multipart/form-data
      description: Загружает PDF, сохраняет метаданные и запускает фоновую обработку.
      parameters:
      - description: PDF-файл
        in: formData
        name: file
        required: true
        type: file
      - description: Название документа
        in: formData
        name: title
        required: true
        type: string
      - description: Авторы
        in: formData
        name: authors
        type: string
      - description: Год публикации
        in: formData
        name: year
        type: string
      - description: Категория
        in: formData
        name: category
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.UploadResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Загрузка PDF
      tags:
      - documents
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"